# openapi-appengine.yaml
swagger: "2.0"
...
info:
  description: "Dantion Endpoints API."
  title: "Endpoints Dantion"
  version: "1.0.0"
host: "dangerdetection.appspot.com"
...
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - "https"
  - "http"

# security:
#   - api_key: []
securityDefinitions:
  # api_key:
  #   type: "apiKey"
  #   name: "key"
  #   in: "query"
  Bearer:
    type: "apiKey"
    name: "Authorization"
    in: "header"
paths:
...
  "/admins/userRole":
    patch:
      tags:
        - "Admin"
      security:
        - Bearer: []
      description: "Update User Role"
      operationId: "UserRole"
      produces:
        - "application/json"
      parameters:
        - in: body
          name: Admin Update User Role
          description: Request body for update user role.
          schema:
            $ref: "#/definitions/AdminUpdateUserRoleBody"
      responses:
        201:
          description: "Success update user role"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
  "/users/{id}":
    get:
      tags:
        - "Users"
      security:
        - Bearer: []
      description: "Get User Using ID"
      operationId: "Userid"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of user"
          required: true
          type: "string"
      responses:
        201:
          description: "success get detail User"
          schema:
            $ref: "#/definitions/UserSuccessResponse"
        400:
          description: "failed get detail User"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
  "/detections/{id}":
    get:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Get Detection Using ID"
      operationId: "detectionId"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of detection"
          required: true
          type: "string"
      responses:
        201:
          description: "success get detail detection"
          schema:
            $ref: "#/definitions/DetectionSuccessResponse"
        400:
          description: "failed get detail detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
    delete:
      tags:
        - "Detections"
      security:
        - Bearer: []
      description: "Delete Detection ID"
      operationId: "deleteDetection"
      produces:
        - "application/json"
      parameters:
        - name: id
          in: path
          description: "ID of detection"
          required: true
          type: "string"
      responses:
        201:
          description: "success delete detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
        400:
          description: "failed delete detection"
          schema:
            $ref: "#/definitions/StatusMessageResponse"
...
definitions:
  StatusMessageResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      message:
        type: "string"
...
  AdminUpdateUserRoleBody:
    type: "object"
    properties:
      userId:
        type: "string"
      role:
        type: "string"
        enum:
          - "polisi"
          - "ambulan"
          - "damkar"
  User:
    type: "object"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      address:
        type: "string"
      number:
        type: "string"
      parentNumber:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
        format: "password"
      role:
        type: "string"
        enum:
          - "admin"
          - "polisi"
          - "ambulan"
          - "damkar"
      photo:
        type: "string"
        format: "binary"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
...
  UserSuccessResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      message:
        type: "string"
      user:
        $ref: "#/definitions/User"
...
  Detection:
    type: "object"
    properties:
      id:
        type: "string"
      lat:
        type: "number"
        format: "double"
      lon:
        type: "number"
        format: "double"
      recordUrl:
        type: "string"
        format: "binary"
      type:
        type: "string"
        enum:
          - "kejahatan"
          - "kecelakaan"
          - "kebakaran"
      status:
        type: "string"
        enum:
          - "invalid"
          - "valid"
          - "selesai"
          - "tolak"
      userId:
        type: "string"
      validatorId:
        type: "string"
      city:
        type: "string"
      createdAt:
        type: "string"
        format: "date-time"
      updatedAt:
        type: "string"
        format: "date-time"
  DetectionSuccessResponse:
    type: "object"
    properties:
      error:
        type: "boolean"
      message:
        type: "string"
      detections:
        $ref: "#/definitions/Detection"
...